refactor: rebuild test setup
