--- examples-groovy/spock13/build.gradle
@@ -1,16 +1,0 @@
-plugins {
-    id 'groovy'
-}
-
-dependencies {
-    implementation platform(project(':dependencies'))
-    implementation platform('org.codehaus.groovy:groovy-bom:2.5.15')
-    implementation 'org.codehaus.groovy:groovy'
-
-    testImplementation platform(project(':dependencies'))
-    testImplementation platform('org.spockframework:spock-bom:1.3-groovy-2.5')
-    testImplementation project(':spock-deepmock')
-    testImplementation 'org.spockframework:spock-core'
-    testImplementation 'net.bytebuddy:byte-buddy'
-    testImplementation 'org.objenesis:objenesis'
-}
