--- examples-groovy/spock13/src/main/groovy/io/github/joke/test/Nested.groovy
@@ -1,12 +1,0 @@
-package io.github.joke.test
-
-
-import groovy.transform.CompileStatic
-import groovy.transform.TupleConstructor
-
-@CompileStatic
-@TupleConstructor
-class Nested {
-    String name
-    Nested child
-}
