--- examples-groovy/spock21/build.gradle
@@ -1,20 +1,0 @@
-plugins {
-    id 'groovy'
-}
-
-dependencies {
-    implementation platform(project(':dependencies'))
-    implementation platform('org.codehaus.groovy:groovy-bom:3.0.9')
-    implementation 'org.codehaus.groovy:groovy'
-
-    testImplementation platform(project(':dependencies'))
-    testImplementation platform('org.spockframework:spock-bom:2.1-groovy-3.0')
-    testImplementation project(':spock-deepmock')
-    testImplementation 'org.spockframework:spock-core'
-    testImplementation 'net.bytebuddy:byte-buddy'
-    testImplementation 'org.objenesis:objenesis'
-}
-
-test {
-    useJUnitPlatform()
-}
