--- examples-groovy/spock21/src/main/groovy/io/github/joke/test/Caller.groovy
@@ -1,14 +1,0 @@
-package io.github.joke.test
-
-import groovy.transform.CompileStatic
-import groovy.transform.TupleConstructor
-
-@CompileStatic
-@TupleConstructor
-class Caller {
-    Nested nested
-
-    def call() {
-        nested.child.child.name
-    }
-}
