--- examples-java/spock13/build.gradle
@@ -1,19 +1,0 @@
-plugins {
-    id 'java'
-    id 'groovy'
-}
-
-dependencies {
-    annotationProcessor platform(project(':dependencies'))
-    annotationProcessor 'org.projectlombok:lombok'
-
-    implementation platform(project(':dependencies'))
-    implementation 'org.projectlombok:lombok'
-
-    testImplementation platform('org.codehaus.groovy:groovy-bom:2.5.15')
-    testImplementation platform('org.spockframework:spock-bom:1.3-groovy-2.5')
-    testImplementation project(':spock-deepmock')
-    testImplementation 'org.spockframework:spock-core'
-    testImplementation 'net.bytebuddy:byte-buddy'
-    testImplementation 'org.objenesis:objenesis'
-}
