--- examples-java/spock20/src/main/java/io/github/joke/test/Nested.java
@@ -1,9 +1,0 @@
-package io.github.joke.test;
-
-import lombok.Data;
-
-@Data
-public class Nested {
-    private final String name;
-    private final Nested child;
-}
