--- examples-java/spock21/build.gradle
@@ -1,23 +1,0 @@
-plugins {
-    id 'java'
-    id 'groovy'
-}
-
-dependencies {
-    annotationProcessor platform(project(':dependencies'))
-    annotationProcessor 'org.projectlombok:lombok'
-
-    implementation platform(project(':dependencies'))
-    implementation 'org.projectlombok:lombok'
-
-    testImplementation platform('org.codehaus.groovy:groovy-bom:3.0.9')
-    testImplementation platform('org.spockframework:spock-bom:2.1-groovy-3.0')
-    testImplementation project(':spock-deepmock')
-    testImplementation 'org.spockframework:spock-core'
-    testImplementation 'net.bytebuddy:byte-buddy'
-    testImplementation 'org.objenesis:objenesis'
-}
-
-test {
-    useJUnitPlatform()
-}
