--- examples-java/spock21/src/main/java/io/github/joke/test/Caller.java
@@ -1,12 +1,0 @@
-package io.github.joke.test;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class Caller {
-    private final Nested nested;
-
-    public String call() {
-        return nested.getChild().getChild().getName();
-    }
-}
