+++ examples/src/main/groovy/io/github/joke/test/GroovyCaller.groovy
@@ -1,0 +1,14 @@
+package io.github.joke.test
+
+import groovy.transform.CompileStatic
+import groovy.transform.TupleConstructor
+
+@CompileStatic
+@TupleConstructor
+class GroovyCaller {
+    GroovyNested nested
+
+    def call() {
+        nested.child.child.name
+    }
+}
