+++ examples/src/main/groovy/io/github/joke/test/GroovyNested.groovy
@@ -1,0 +1,12 @@
+package io.github.joke.test
+
+
+import groovy.transform.CompileStatic
+import groovy.transform.TupleConstructor
+
+@CompileStatic
+@TupleConstructor
+class GroovyNested {
+    String name
+    GroovyNested child
+}
