+++ examples/src/main/java/io/github/joke/test/Caller.java
@@ -1,0 +1,13 @@
+package io.github.joke.test;
+
+public class Caller {
+    private final Nested nested;
+
+    public Caller(final Nested nested) {
+        this.nested = nested;
+    }
+
+    public String call() {
+        return nested.getChild().getChild().getName();
+    }
+}
