+++ examples/src/main/java/io/github/joke/test/Nested.java
@@ -1,0 +1,19 @@
+package io.github.joke.test;
+
+public class Nested {
+    private final String name;
+    private final Nested child;
+
+    public Nested(final String name, final Nested child) {
+        this.name = name;
+        this.child = child;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public Nested getChild() {
+        return child;
+    }
+}
