+++ examples/src/test/groovy/io/github/joke/test/MockTest.groovy
